fix login&register fahrezi
